Reorganize test logic to match reducer refactoring.
